+++ java/project/srv/src/main/java/com/github/cap/js/community/sapafcsdk/scheduling/common/JobSchedulingException.java
@@ -6,7 +6,7 @@ import com.sap.cds.services.ServiceException;
 
 public class JobSchedulingException extends ServiceException {
 
-  private String code;
+  private final String code;
 
   public JobSchedulingException(ErrorStatus errorStatus, String messageOrKey, Object... args) {
     super(errorStatus, messageOrKey, args);