+++ test/scheduling/processing.test.js
@@ -391,6 +391,7 @@ describe("Processing Service", () => {
   it("updateJob - results - arraybuffer", async () => {
     await expect(processingService.updateJob(ID, JobStatus.running)).resolves.not.toThrow();
     await processQueue("sapafcsdk.scheduling.SchedulingProcessingService");
+    await clearEventQueue();
     await expect(
       processingService.updateJob(ID, JobStatus.completed, [
         {
@@ -403,7 +404,15 @@ describe("Processing Service", () => {
       ]),
     ).resolves.not.toThrow();
     await processQueue("sapafcsdk.scheduling.SchedulingProcessingService");
-    expect(log.output).toEqual(expect.stringMatching(/ASSERT_DATA_TYPE.*LargeBinary {.*type: 'cds.LargeBinary'.*}/s));
+    const entry = await eventQueueEntry();
+    expect(entry).toBeDefined();
+    expect(entry.status).toBe(3);
+    expect(JSON.parse(entry.error)).toMatchObject({
+      message: "ASSERT_DATA_TYPE",
+      name: "Error",
+      stack: "ASSERT_DATA_TYPE",
+      target: "data",
+    });
   });
 
   it("updateJob - processJobUpdate", async () => {
@@ -487,8 +496,15 @@ describe("Processing Service", () => {
       const entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/jobNotFound.*XXX/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: ["XXX"],
+        code: "jobNotFound",
+        message: "jobNotFound",
+        name: "jobNotFound",
+        numericSeverity: 4,
+        severity: "E",
+        status: 404,
+      });
       await clearEventQueue();
     });
 
@@ -498,8 +514,15 @@ describe("Processing Service", () => {
       let entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/jobNotFound.*XXX/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: ["XXX"],
+        code: "jobNotFound",
+        message: "jobNotFound",
+        name: "jobNotFound",
+        numericSeverity: 4,
+        severity: "E",
+        status: 404,
+      });
       await clearEventQueue();
 
       await expect(processingService.updateJob(ID)).resolves.not.toThrow();
@@ -507,8 +530,15 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/statusValueMissing/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: [],
+        code: "statusValueMissing",
+        message: "statusValueMissing",
+        name: "statusValueMissing",
+        numericSeverity: 4,
+        severity: "E",
+        status: 400,
+      });
       await clearEventQueue();
 
       await expect(processingService.updateJob(ID, "XXX")).resolves.not.toThrow();
@@ -516,8 +546,15 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/invalidJobStatus.*XXX/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: ["XXX"],
+        code: "invalidJobStatus",
+        message: "invalidJobStatus",
+        name: "invalidJobStatus",
+        numericSeverity: 4,
+        severity: "E",
+        status: 400,
+      });
       await clearEventQueue();
 
       await expect(processingService.updateJob(ID, JobStatus.completed)).resolves.not.toThrow();
@@ -525,8 +562,15 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/statusTransitionNotAllowed.*requested.*completed/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: ["requested", "completed"],
+        code: "statusTransitionNotAllowed",
+        message: "statusTransitionNotAllowed",
+        name: "statusTransitionNotAllowed",
+        numericSeverity: 4,
+        severity: "E",
+        status: 400,
+      });
       await clearEventQueue();
     });
 
@@ -536,8 +580,12 @@ describe("Processing Service", () => {
       let entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/message: 'ASSERT_ARRAY', target: 'results'/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        name: "Error",
+        message: "ASSERT_ARRAY",
+        stack: "ASSERT_ARRAY",
+        target: "results",
+      });
       await clearEventQueue();
 
       await expect(processingService.updateJob(ID, JobStatus.running, [{}])).resolves.not.toThrow();
@@ -545,8 +593,15 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/resultNameMissing/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        name: "resultNameMissing",
+        message: "resultNameMissing",
+        code: "resultNameMissing",
+        args: [],
+        status: 400,
+        severity: "E",
+        numericSeverity: 4,
+      });
       await clearEventQueue();
 
       await expect(processingService.updateJob(ID, JobStatus.running, [{ name: "Link" }])).resolves.not.toThrow();
@@ -554,8 +609,15 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/resultTypeMissing/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: [],
+        code: "resultTypeMissing",
+        message: "resultTypeMissing",
+        name: "resultTypeMissing",
+        numericSeverity: 4,
+        severity: "E",
+        status: 400,
+      });
       await clearEventQueue();
 
       await expect(
@@ -565,8 +627,15 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/invalidResultType.*'X'/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: ["X"],
+        code: "invalidResultType",
+        message: "invalidResultType",
+        name: "invalidResultType",
+        numericSeverity: 4,
+        severity: "E",
+        status: 400,
+      });
       await clearEventQueue();
 
       await expect(
@@ -581,8 +650,15 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/linkMissing.*link/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: ["link"],
+        code: "linkMissing",
+        message: "linkMissing",
+        name: "linkMissing",
+        numericSeverity: 4,
+        severity: "E",
+        status: 400,
+      });
       await clearEventQueue();
 
       await expect(
@@ -599,8 +675,15 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/mimeTypeNotAllowed.*link/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: ["link"],
+        code: "mimeTypeNotAllowed",
+        message: "mimeTypeNotAllowed",
+        name: "mimeTypeNotAllowed",
+        numericSeverity: 4,
+        severity: "E",
+        status: 400,
+      });
       await clearEventQueue();
 
       await expect(
@@ -617,8 +700,15 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/filenameNotAllowed.*link/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: ["link"],
+        code: "filenameNotAllowed",
+        message: "filenameNotAllowed",
+        name: "filenameNotAllowed",
+        numericSeverity: 4,
+        severity: "E",
+        status: 400,
+      });
       await clearEventQueue();
 
       await expect(
@@ -635,8 +725,15 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/dataNotAllowed.*link/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: ["link"],
+        code: "dataNotAllowed",
+        message: "dataNotAllowed",
+        name: "dataNotAllowed",
+        numericSeverity: 4,
+        severity: "E",
+        status: 400,
+      });
       await clearEventQueue();
 
       await expect(
@@ -653,8 +750,15 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/messagesNotAllowed.*link/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: ["link"],
+        code: "messagesNotAllowed",
+        message: "messagesNotAllowed",
+        name: "messagesNotAllowed",
+        numericSeverity: 4,
+        severity: "E",
+        status: 400,
+      });
       await clearEventQueue();
 
       await expect(processingService.updateJob(ID, JobStatus.running, [{ name: "Data" }])).resolves.not.toThrow();
@@ -662,8 +766,15 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/resultTypeMissing/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: [],
+        code: "resultTypeMissing",
+        message: "resultTypeMissing",
+        name: "resultTypeMissing",
+        numericSeverity: 4,
+        severity: "E",
+        status: 400,
+      });
       await clearEventQueue();
 
       await expect(
@@ -678,8 +789,15 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/mimeTypeMissing.*data/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: ["data"],
+        code: "mimeTypeMissing",
+        message: "mimeTypeMissing",
+        name: "mimeTypeMissing",
+        numericSeverity: 4,
+        severity: "E",
+        status: 400,
+      });
       await clearEventQueue();
 
       await expect(
@@ -695,8 +813,15 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/filenameMissing.*data/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: ["data"],
+        code: "filenameMissing",
+        message: "filenameMissing",
+        name: "filenameMissing",
+        numericSeverity: 4,
+        severity: "E",
+        status: 400,
+      });
       await clearEventQueue();
 
       await expect(
@@ -713,8 +838,15 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/dataMissing.*data/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: ["data"],
+        code: "dataMissing",
+        message: "dataMissing",
+        name: "dataMissing",
+        numericSeverity: 4,
+        severity: "E",
+        status: 400,
+      });
       await clearEventQueue();
 
       await expect(
@@ -733,8 +865,15 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/linkNotAllowed.*data/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: ["data"],
+        code: "linkNotAllowed",
+        message: "linkNotAllowed",
+        name: "linkNotAllowed",
+        numericSeverity: 4,
+        severity: "E",
+        status: 400,
+      });
       await clearEventQueue();
 
       await expect(
@@ -753,8 +892,15 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/messagesNotAllowed.*data/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: ["data"],
+        code: "messagesNotAllowed",
+        message: "messagesNotAllowed",
+        name: "messagesNotAllowed",
+        numericSeverity: 4,
+        severity: "E",
+        status: 400,
+      });
       await clearEventQueue();
 
       await expect(
@@ -769,8 +915,15 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/messagesMissing.*message/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: ["message"],
+        code: "messagesMissing",
+        message: "messagesMissing",
+        name: "messagesMissing",
+        numericSeverity: 4,
+        severity: "E",
+        status: 400,
+      });
       await clearEventQueue();
 
       await expect(
@@ -786,8 +939,12 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/ASSERT_ARRAY', target: 'messages'/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        message: "ASSERT_ARRAY",
+        name: "Error",
+        stack: "ASSERT_ARRAY",
+        target: "messages",
+      });
       await clearEventQueue();
 
       await expect(
@@ -803,8 +960,15 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/messagesMissing.*message/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: ["message"],
+        code: "messagesMissing",
+        message: "messagesMissing",
+        name: "messagesMissing",
+        numericSeverity: 4,
+        severity: "E",
+        status: 400,
+      });
       await clearEventQueue();
 
       await expect(
@@ -820,8 +984,15 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/codeMissing/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: [],
+        code: "codeMissing",
+        message: "codeMissing",
+        name: "codeMissing",
+        numericSeverity: 4,
+        severity: "E",
+        status: 400,
+      });
       await clearEventQueue();
 
       await expect(
@@ -841,8 +1012,15 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/textMissing/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: [],
+        code: "textMissing",
+        message: "textMissing",
+        name: "textMissing",
+        numericSeverity: 4,
+        severity: "E",
+        status: 400,
+      });
       await clearEventQueue();
 
       await expect(
@@ -863,8 +1041,15 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/severityMissing/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: [],
+        code: "severityMissing",
+        message: "severityMissing",
+        name: "severityMissing",
+        numericSeverity: 4,
+        severity: "E",
+        status: 400,
+      });
       await clearEventQueue();
 
       await expect(
@@ -886,8 +1071,15 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/invalidMessageSeverity.*'X'/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: ["X"],
+        code: "invalidMessageSeverity",
+        message: "invalidMessageSeverity",
+        name: "invalidMessageSeverity",
+        numericSeverity: 4,
+        severity: "E",
+        status: 400,
+      });
       await clearEventQueue();
 
       await expect(
@@ -910,8 +1102,18 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/ASSERT_DATA_TYPE.*Timestamp { type: 'cds.Timestamp' }/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: [
+          "xxx",
+          {
+            type: "cds.Timestamp",
+          },
+        ],
+        message: "ASSERT_DATA_TYPE",
+        name: "Error",
+        stack: "ASSERT_DATA_TYPE",
+        target: "createdAt",
+      });
       await clearEventQueue();
 
       await expect(
@@ -934,8 +1136,15 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/linkNotAllowed/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: ["message"],
+        code: "linkNotAllowed",
+        message: "linkNotAllowed",
+        name: "linkNotAllowed",
+        numericSeverity: 4,
+        severity: "E",
+        status: 400,
+      });
       await clearEventQueue();
 
       await expect(
@@ -958,8 +1167,15 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/mimeTypeNotAllowed/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: ["message"],
+        code: "mimeTypeNotAllowed",
+        message: "mimeTypeNotAllowed",
+        name: "mimeTypeNotAllowed",
+        numericSeverity: 4,
+        severity: "E",
+        status: 400,
+      });
       await clearEventQueue();
 
       await expect(
@@ -982,8 +1198,15 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/filenameNotAllowed/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: ["message"],
+        code: "filenameNotAllowed",
+        message: "filenameNotAllowed",
+        name: "filenameNotAllowed",
+        numericSeverity: 4,
+        severity: "E",
+        status: 400,
+      });
       await clearEventQueue();
 
       await expect(
@@ -1006,8 +1229,15 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/dataNotAllowed/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: ["message"],
+        code: "dataNotAllowed",
+        message: "dataNotAllowed",
+        name: "dataNotAllowed",
+        numericSeverity: 4,
+        severity: "E",
+        status: 400,
+      });
       await clearEventQueue();
 
       await expect(
@@ -1029,7 +1259,16 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/localeMissing/s));
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: [],
+        code: "localeMissing",
+        message: "localeMissing",
+        name: "localeMissing",
+        numericSeverity: 4,
+        severity: "E",
+        status: 400,
+      });
+      await clearEventQueue();
 
       await expect(
         processingService.updateJob(ID, JobStatus.running, [
@@ -1050,7 +1289,15 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/invalidLocale.*xx/s));
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: ["xx"],
+        code: "invalidLocale",
+        message: "invalidLocale",
+        name: "invalidLocale",
+        numericSeverity: 4,
+        severity: "E",
+        status: 400,
+      });
     });
 
     it("cancelJob - status", async () => {
@@ -1059,21 +1306,34 @@ describe("Processing Service", () => {
       let entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/jobNotFound.*XXX/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: ["XXX"],
+        code: "jobNotFound",
+        message: "jobNotFound",
+        name: "jobNotFound",
+        numericSeverity: 4,
+        severity: "E",
+        status: 404,
+      });
       await clearEventQueue();
 
       await expect(processingService.updateJob(ID, "running")).resolves.not.toThrow();
       await processQueue("sapafcsdk.scheduling.SchedulingProcessingService");
-      log.clear();
       await clearEventQueue();
       await expect(processingService.cancelJob(ID)).resolves.not.toThrow();
       await processQueue("sapafcsdk.scheduling.SchedulingProcessingService");
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/statusTransitionNotAllowed.*running.*canceled/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: ["running", "canceled"],
+        code: "statusTransitionNotAllowed",
+        message: "statusTransitionNotAllowed",
+        name: "statusTransitionNotAllowed",
+        numericSeverity: 4,
+        severity: "E",
+        status: 400,
+      });
       await clearEventQueue();
     });
 
@@ -1107,8 +1367,15 @@ describe("Processing Service", () => {
       entry = await eventQueueEntry();
       expect(entry).toBeDefined();
       expect(entry.status).toBe(3);
-      expect(log.output).toEqual(expect.stringMatching(/statusTransitionNotAllowed.*completed.*canceled/s));
-      log.clear();
+      expect(JSON.parse(entry.error)).toMatchObject({
+        args: ["completed", "canceled"],
+        code: "statusTransitionNotAllowed",
+        message: "statusTransitionNotAllowed",
+        name: "statusTransitionNotAllowed",
+        numericSeverity: 4,
+        severity: "E",
+        status: 400,
+      });
       await clearEventQueue();
     });